movies/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.92.0
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
grpcio-tools==1.51.3
h11==0.14.0
httptools==0.5.0
idna==3.4
protobuf==4.22.0
pydantic==1.10.5
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
